feat(stemmer,sentences): add WASM bindings via core split
stemmer: extracts Porter/Snowball + tokenization logic into
`_stemmer-core`; the napi crate wraps it. WASM exposes the same
`Stemmer` class (constructor + stemMany/stemBuffer/tokenizeAndStem/
tokenizeAndStemToBuffer/language getter) plus the `stemOnce` free
function. Option struct flows through serde-wasm-bindgen.

sentences: extracts the rule-based SBD engine and per-language
abbreviation tables into `_sentences-core`; the napi crate now
contains only the FFI surface. WASM exposes `split`, `splitToOffsets`
(returning Uint8Array of u32 LE offset pairs), and `splitBatch`.
