+++ crates/_sentences-core/src/abbreviations.rs
