+++ crates/stemmer/package.json
@@ -4,10 +4,28 @@
   "description": "Porter/Snowball stemmer — batch-only, 18 languages, integrates with @amigo-labs/bm25.",
   "main": "index.js",
   "types": "index.d.ts",
+  "browser": "./wasm/pkg/amigo_stemmer_wasm.js",
+  "exports": {
+    ".": {
+      "browser": {
+        "types": "./wasm/pkg/amigo_stemmer_wasm.d.ts",
+        "default": "./wasm/pkg/amigo_stemmer_wasm.js"
+      },
+      "default": {
+        "types": "./index.d.ts",
+        "default": "./index.js"
+      }
+    }
+  },
   "files": [
     "index.js",
     "index.d.ts",
-    "README.md"
+    "README.md",
+    "wasm/pkg/amigo_stemmer_wasm.js",
+    "wasm/pkg/amigo_stemmer_wasm_bg.js",
+    "wasm/pkg/amigo_stemmer_wasm_bg.wasm",
+    "wasm/pkg/amigo_stemmer_wasm.d.ts",
+    "wasm/pkg/amigo_stemmer_wasm_bg.wasm.d.ts"
   ],
   "napi": {
     "binaryName": "stemmer",
@@ -24,9 +42,12 @@
     "artifacts": "napi artifacts",
     "build": "napi build --platform --release",
     "build:debug": "napi build --platform",
-    "prepublishOnly": "napi pre-publish --skip-optional-publish --no-gh-release",
+    "build:wasm": "cd wasm && wasm-pack build --target bundler --release --out-dir pkg",
+    "build:all": "pnpm build && pnpm build:wasm",
+    "prepublishOnly": "pnpm build:wasm && napi pre-publish --skip-optional-publish --no-gh-release",
     "test": "vitest run __test__",
     "test:conformance": "vitest run __conformance__",
+    "test:wasm": "cd wasm && wasm-pack test --node",
     "test:all": "vitest run",
     "bench": "vitest bench"
   },
@@ -64,6 +85,7 @@
     "vsJs": "7.1-8.3x",
     "parity": "100%",
     "status": "Alternative",
+    "targets": ["node", "browser"],
     "competitors": [
       "natural@8"
     ]
